Updated the proxies snippet as the config was missing for the AzureIdentityAuthenticationProvider
Updated the proxies snippet as the config was missing for the AzureIdentityAuthenticationProvider, also the proxy config differs between the AzureIdentityAuthenticationProvider and the httpx AsyncClient as the first one uses the requests libary and the second one use httpx so the keys are different.  Also if you only pass the proxy config to the AsyncClient, the authentication will fail as it doesn't have the proxy config as it uses another client internally.